Date subcommands (#2383)
* refactored date command

* format files
