+++ crates/nu-cli/src/cli.rs
@@ -261,7 +261,9 @@ pub fn create_default_context(
             whole_stream_command(Save),
             whole_stream_command(Touch),
             whole_stream_command(Cpy),
-            whole_stream_command(Date),
+            whole_stream_command(DateNow),
+            whole_stream_command(DateUTC),
+            whole_stream_command(DateFormat),
             whole_stream_command(Cal),
             whole_stream_command(Mkdir),
             whole_stream_command(Mv),
